requirements.txt
fastapi==0.87.0
pytest==7.2.0
python-dotenv==0.21.0
uvicorn==0.19.0
requests==2.28.1
gunicorn==20.1.0